refactor: update imports and references for new sync service structure
- Update service manager to use new sync service paths
- Update main.go to import from new stagedstreamsync location
- Update node configuration and syncing logic
- Update Rosetta configurations
- Update test deployment scripts
- Update Docker proto image build workflow
