+++ .github/workflows/build-docker-proto-image.yaml
@@ -27,8 +27,8 @@ jobs:
       - name: Build and push
         uses: docker/build-push-action@v6
         with:
-          context: ./harmony/api/service/legacysync/downloader
-          file: ./harmony/api/service/legacysync/downloader/Proto.Dockerfile
+          context: ./harmony/api/service/synchronize/legacysync/downloader
+          file: ./harmony/api/service/synchronize/legacysync/downloader/Proto.Dockerfile
           push: true
           platforms: linux/amd64,linux/arm64
           tags: |
