+++ api/service/manager.go
@@ -23,7 +23,6 @@ const (
 	Prometheus
 	Synchronize
 	CrosslinkSending
-	StagedStreamSync
 )
 
 func (t Type) String() string {
@@ -46,8 +45,6 @@ func (t Type) String() string {
 		return "Synchronize"
 	case CrosslinkSending:
 		return "CrosslinkSending"
-	case StagedStreamSync:
-		return "StagedStreamSync"
 	default:
 		return "Unknown"
 	}
