+++ rosetta/infra/harmony-pstn.conf
@@ -107,12 +107,11 @@ Version = "2.5.13"
   DiscHardLowCap = 6
   DiscHighCap = 128
   DiscSoftLowCap = 8
-  Downloader = false
+  Client = false
   Enabled = false
   InitStreams = 8
   MaxAdvertiseWaitTime = 30
   MinPeers = 2
-  StagedSync = false
 
   [Sync.StagedSyncCfg]
     DoubleCheckBlockHashes = false
@@ -121,7 +120,6 @@ Version = "2.5.13"
     MaxBackgroundBlocks = 0
     MaxBlocksPerSyncCycle = 0
     MaxMemSyncCycleSize = 0
-    TurboMode = false
     UseMemDB = false
     VerifyAllSig = false
     VerifyHeaderBatchSize = 0
